+++ rector.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 use Rector\CodeQuality\Rector\Class_\InlineConstructorDefaultToPropertyRector;
 use Rector\Config\RectorConfig;
 use Rector\Php80\Rector\Class_\ClassPropertyAssignToConstructorPromotionRector;
+use Rector\PHPUnit\Set\PHPUnitSetList;
 use Rector\Set\ValueObject\LevelSetList;
 use Rector\Symfony\Set\FOSRestSetList;
 use Rector\Symfony\Set\JMSSetList;
@@ -45,6 +46,8 @@ return static function (RectorConfig $rectorConfig): void {
         FOSRestSetList::ANNOTATIONS_TO_ATTRIBUTES,
         JMSSetList::ANNOTATIONS_TO_ATTRIBUTES,
         SensiolabsSetList::ANNOTATIONS_TO_ATTRIBUTES,
+
+        PHPUnitSetList::PHPUNIT_100,
 
         LevelSetList::UP_TO_PHP_81,
     ]);